+++ server/api/communities.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const { Community, User, CommunityTask, Task } = require('../db/models')
+const { Community, User, CommunityTask, Task, TaskItem } = require('../db/models')
 const sendEmail = require('../mailer')
 module.exports = router
 
@@ -44,6 +44,38 @@ router.post('/:communityId/inviteUsers', (req, res, next) => {
     emails.forEach(email => {
       sendEmail(email, user, req.community, 'invite')
     })
+  }
+  catch (err) {
+    next(err)
+  }
+})
+
+
+router.put('/:communityId/tasks', async (req, res, next) => {
+  try {
+    const tasks = req.body
+    
+    const communityTaskPromises = tasks.map(task => {
+      CommunityTask.update({
+        value: task.value
+      }, {
+        where: {
+          id: task.id,
+        }
+      })
+    })
+    const taskItemPromises = tasks.map(task => {
+      TaskItem.findOrCreate({
+        where: {
+          taskId: task.taskId,
+          communityId: req.params.communityId,
+          completed: null,
+        }
+      })
+    })
+    await Promise.all(communityTaskPromises)
+    await Promise.all(taskItemPromises)
+    res.json(201)
   }
   catch (err) {
     next(err)
