+++ server/api/community-tasks.js
@@ -32,34 +32,3 @@ router.post('/:communityId', async (req, res, next) => {
     next(err)
   }
 })
-
-router.put('/frequencies/:communityId', async (req, res, next) => {
-  try {
-    const tasks = req.body
-    
-    const communityTaskPromises = tasks.map(task => {
-      CommunityTask.update({
-        value: task.value
-      }, {
-        where: {
-          id: task.id,
-        }
-      })
-    })
-    const taskItemPromises = tasks.map(task => {
-      TaskItem.findOrCreate({
-        where: {
-          taskId: task.taskId,
-          communityId: req.params.communityId,
-          completed: null,
-        }
-      })
-    })
-    await Promise.all(communityTaskPromises)
-    await Promise.all(taskItemPromises)
-    res.json(201)
-  }
-  catch (err) {
-    next(err)
-  }
-})
